move target_os to platform
