+++ base/base.cpp
@@ -4,7 +4,7 @@
 #include "base/exception.hpp"
 #include "base/src_point.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <iostream>
