+++ base/condition.cpp
@@ -1,7 +1,7 @@
 #include "base/condition.hpp"
 #include "base/mutex.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <pthread.h>
