+++ base/string_utils.cpp
@@ -1,7 +1,7 @@
 #include "base/assert.hpp"
 #include "base/string_utils.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <algorithm>
 #include <cmath>
