+++ base/thread.hpp
@@ -4,7 +4,7 @@
 #include "base/cancellable.hpp"
 #include "base/macros.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <cstdint>
 #include <functional>
