+++ base/timer.cpp
@@ -6,7 +6,7 @@
 #include "base/macros.hpp"
 #include "base/timegm.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <algorithm>
 #include <array>
