+++ indexer/drawing_rules.hpp
@@ -7,7 +7,7 @@
 #include "base/base.hpp"
 #include "base/buffer_vector.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <array>
 #include <iostream>
