+++ indexer/drules_include.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include "indexer/drules_struct.pb.h"
