+++ indexer/feature_meta.cpp
@@ -1,6 +1,6 @@
 #include "indexer/feature_meta.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <sstream>
