+++ indexer/indexer_tests/feature_metadata_test.cpp
@@ -5,7 +5,7 @@
 #include "coding/reader.hpp"
 #include "coding/writer.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <map>
 #include <string>
