+++ platform/platform_std.cpp
@@ -7,7 +7,7 @@
 
 #include "base/logging.hpp"
 
-#include "std/target_os.hpp"
+#include "platform/target_os.hpp"
 
 #include <algorithm>
 #include <future>
